After button press, poll until state matches expected
This needs a bunch of error handling added to it (what if the button press doesn't go through for some reason? etc.) but proves the concept in the ideal case. Now the page will continue to refresh every 5 seconds after a button press until the door is moved to the state it's expected to be in.

This also resets the node_helper 5-minute auto-update timer every time an update is queued so that frontend updates/door toggles don't overlap with backend update intervals.