Update appearance, support multiple devices
This changes the appearance such that the button is a little closer to the desired end result. The device's name and current state are printed on the button which is large enough to be easily pressed on a touch screen.

This also updates from the single-device expectation to support multiple devices being present at once.