Get rid of unnecessary getDoorState() step
The myq-api library just calls into getDevices() again and filters the response down to just the device and attribute requested, so we can skip this and handle it ourselves without the extra hit to the API and the delay that request incurs. All necessary state information is included on the device object.

The only minor bummer is that we were previously able to skip a full DOM update when the state hadn't changed, but realistically that pattern was dicey anyway since if there is more than one device being monitored, we'd potentially need to push an update for other device(s) anyway.