Switch to a working MyQ API library
The myq-api lib I was using is still not updated to the new API, and as
a result doesn't work. This lib is recent and seems to be powering an
active project.

I'm sure this isn't exactly how it needs to be done long-term, but this
gets things functional for me right now and I can iterate on it as I go.